+++ tests/dynamic_data/users_roles/test_current_user.py
@@ -22,6 +22,14 @@ def test_patch():
         'code': 30,
         'message': 'Unauthorized access',
     }, id='unauthorized'),
+    pytest.param({
+        'username': 'test_patch_current_user_username_fail_'.ljust(60, 'x'),
+        'actor_name': 'test_patch_current_user_username_fail_'.ljust(60, 'x')
+    }, {
+        'http_code': 403,
+        'code': 56,
+        'message': 'Username or workspace name is too long. Maximum is 59 characters.',
+    }, id='too_long'),
 ])
 def test_patch_current_user_raises(params, exp_exception):
     with pytest.raises(LaymanError) as exc_info: